+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This is the changelog for [Authress SDK](readme.md).
 * Fix service client token generated tokens so that they automatically get the custom domain injected in during using in the SDK.
 * Fix TokenVerifier return type to be a `Promise`.
 * Add the `TokenVerificationError` error and `verifyToken` method directly to the Authress Client interface.
+* Add the `ApiError` error type
 
 ## 1.3 ##
 * Add new `Pagination` type which pagination `next.cursor` to enable paging through resources.